优化删除远端服务器上的好友关系-MainActivity-onContactDeleted方法里

+++ src/cn/ucai/superwechat/activity/MainActivity.java
@@ -581,34 +581,27 @@ public class MainActivity extends BaseActivity implements EMEventListener {
 		public void onContactDeleted(final List<String> usernameList) {
 			Log.e(TAG,"onContactDeleted,usernameList"+usernameList);
 			// 被删除
-			Map<String, User> localUsers = ((DemoHXSDKHelper)HXSDKHelper.getInstance()).getContactList();
-			List<String> toDelUserName = new ArrayList<String>();
 			String currentUserName = SuperWeChatApplication.getInstance().getUserName();
 			Log.e(TAG,"onContactDeleted,currentUserName"+currentUserName);
-			for (String username : usernameList) {
+			for (final String username : usernameList) {
 				Log.e(TAG,"onContactDeleted,username"+username);
-				localUsers.remove(username);
-				toDelUserName.add(username);
-				userDao.deleteContact(username);
-				inviteMessgeDao.deleteMessage(username);
-			}
-
-			for (final String name:toDelUserName){
 				final OkHttpUtils2<Result> utils = new OkHttpUtils2<Result>();
 				utils.setRequestUrl(I.REQUEST_DELETE_CONTACT)
 						.addParam(I.Contact.USER_NAME,currentUserName)
-						.addParam(I.Contact.CU_NAME,name)
+						.addParam(I.Contact.CU_NAME,username)
 						.targetClass(Result.class)
 						.execute(new OkHttpUtils2.OnCompleteListener<Result>() {
 							@Override
 							public void onSuccess(Result result) {
-								Map<String, UserAvatar> userMap = SuperWeChatApplication.getInstance().getUserMap();
-								List<UserAvatar> userList = SuperWeChatApplication.getInstance().getUserList();
-								UserAvatar u = userMap.get(name);
-								userList.remove(u);
-								userMap.remove(name);
-								sendStickyBroadcast(new Intent("update_contact_list"));
-
+								if (result.isRetMsg()) {
+									((DemoHXSDKHelper) HXSDKHelper.getInstance()).getContactList().remove(username);
+									UserAvatar u = SuperWeChatApplication.getInstance().getUserMap().get(username);
+									SuperWeChatApplication.getInstance().getUserList().remove(u);
+									SuperWeChatApplication.getInstance().getUserMap().remove(username);
+									userDao.deleteContact(username);
+									inviteMessgeDao.deleteMessage(username);
+									sendStickyBroadcast(new Intent("update_contact_list"));
+								}
 							}
 
 							@Override
@@ -616,7 +609,6 @@ public class MainActivity extends BaseActivity implements EMEventListener {
 
 							}
 						});
-
 			}
 
 			runOnUiThread(new Runnable() {
